VariationPileup

HapLink.VariationPileup — Type

VariationPileup

Summarizes the basecalls that support a Variation within a set of alignments.

Fields

variation::Variation: The Variation of this pileup entry

depth::UInt: The number of times the position of variation appears in this set of alignments

readpos::Vector{Float64}: The relative positions of variation within each read

quality::Vector{Float64}: The phred quality of variation within each read

strand::Vector{Strand}: Which strand each variation is found on

HapLink.altdepth — Method

altdepth(vp::VariationPileup)

Gets the number of times vp.variation appears

HapLink.depth — Method

depth(vp::VariationPileup)

Gets the number of times the position of vp.variation appears total (variant and wild-type)

HapLink.frequency — Method

frequency(vp::VariationPileup)

Gets the alternate allele frequency of vp.variation

HapLink.pileup — Method

pileup(sam::AbstractPath, ref::AbstractPath) -> Vector{VariationPileup}

Generates a pileup of Variations based on the alignments in sam aligned against ref.

HapLink.quality — Method

quality(vp::VariationPileup)

Gets the phred qualities of vp.variation

HapLink.readpos — Method

readpos(vp::VariationPileup)

Gets the relative positions of vp.variation

HapLink.strand — Method

strand(vp::VariationPileup)

Gets the strands of vp.variation

HapLink.strand_bias — Method

strand_bias(vp::VariationPileup)

Gets the frequency of positive strands that variation appears on relative to all variation reads

HapLink.variation — Method

variation(vp::VariationPileup)

Gets the Variation of vp