haplink consensus

HapLink.consensus — Function

haplink consensus [options] reference variants

Convert variant calls to consensus sequence

Introduction

Generates a consensus sequence based on a reference genome and previously called variants. Will only consider variants with a "PASS" filter to be able to contribute to the consensus. Outputs results in FASTA format.

Arguments

reference: The path to the reference genome. Must be in FASTA format.

variants: The path to the variant calls. Must be in VCF v4 format.

Options

--outfile=<path>: The file to write the consensus sequence to. If left blank, the consensus sequence is written to standard output.

--frequency=<float>: The minimum frequency at which a variant must appear to be considered part of the consensus. Note that HapLink does not support ambigous base calling (e.g. N, R, Y, etc.) at low frequencies unlike many variant callers.

--prefix=<string>: Name of the new sequence. Defaults to using the FASTA identifier of the reference sequence.