haplink sequences

HapLink.sequences — Function

haplink sequences [options] reference haplotypes

Convert haplotype calls into haplotype sequences

Introduction

Takes a YAML file output from haplink haplotypes and converts each called haplotype into its sequence and outputs those sequences in FASTA format. Useful for downstream processing by other tools, but loses all of the metadata and statistical details of each haplotype.

Arguments

reference: path to the reference genome to mutate haplotypes from. Must not be gzipped, but does not need to be indexed (have a sidecar fai file). HapLink only supports single- segment reference genomes: if reference includes more than one sequence, all but the first will be ignored.

haplotypes: path to the haplotype calls file that will be used to mutate the reference sequence to produce haplotype sequences. Must be an output from haplink haplotypes.

Options

--outfile=<path>: The file to write the haplotype sequences to. If left blank, the sequences are written to standard output.

--prefix=<string>: Name of the new sequences. Defaults to using the FASTA identifier of the reference sequence.