haplink variants

HapLink.variants — Function

haplink variants [options] reference bam

Call variants

Introduction

Decides which variations found within an alignment are real, and which are due to random chance. HapLink uses Fisher's Exact Test to determine the statistical significance of sequence variations, and optionally allows for other thresholds to reduce random noise in the variant calling. Outputs a Variant Call Format (VCF) file compliant with VCF v4.

Arguments

reference: path to the reference genome to call variants against in fasta format. Must not be gzipped, but does not need to be indexed (have a sidecar fai file). HapLink only supports single-segment reference genomes: if reference includes more than one sequence, all but the first will be ignored.

bam: alignment file to call variants from. Can be in SAM or BAM format, and does not need to be sorted or indexed, but variant calling speed will increase significantly if using a sorted and indexed (has a sidebar bai file) BAM file.

Options

--outfile=<path>: The file to write variant calls to. If left blank, variant calls are written to standard output.

--significance=<float>: The alpha value for statistical significance of variant calls.

--depth=<int>: Minimum number of times the variation must be observed within the alignment to be called a variant

--quality=<float>: The minimum average basecall quality score for a variation to be called a variant

--frequency=<float>: The minimum proportion of reads that the variation must be observed within compared to all reads covering its position for that variation to be called a variant

--position=<float>: The distance (as a percentage) from the edge of reads that a variation must be observed at to be called a variant

--strandedness=<float>: The maximum proportion of times that a variation can be observed on one strand versus the other to be called a variant. This metric is totally useless on single-stranded sequencing protocols like Oxford Nanopore, but can be useful for combining data between stranded protocols like most Illumina and Pacific Bio.