HapLink

HapLink.HapLink — Module

HapLink

Viral haplotype calling via linkage disequilibrium

Welcome

Howdy! 🤠 And welcome to HapLink! 👋 HapLink is a command-line suite of tools to enable the exploration of viral quasispecies within a single metagenomic sample. Every piece eventually builds up to our viral haplotype caller, which uses linkage disequilibrium on long sequencing reads (💡 think Oxford Nanopore or PacBio HiFi) to identify genetic mutations that are conserved within a single virus particle.

This manual will cover the different ways of using HapLink, starting with a few tutorials before diving into the details of our reference section.

Contents

Consensus Tools

Haplotype Methods

HaplotypeCall

Private API Reference

PseudoRead

Variation Methods

VariationCall

VariationInfo

VariationPileup

haplink consensus

haplink haplotypes

haplink sequences

haplink variants

HapLink

Welcome

Getting started

In the beginning

Bioconda

Comonicon

Kicking the tires

Getting the goods

Spot the difference

The general lay of the land

The star attraction

Haplotypes in the Matrix

But, what does it mean?

Playing well with others

Install the other tools

Get sample data

Align sequences

Call variants

Call haplotypes

Alignment (again, but on a bigger scale)

Phylogeny

For advanced beginners (REPL mode)

Installing HapLink ... into a project

Finding the example files

Pileup time

But is it real?

Rule of the majority

Bringing in the reads

Finding haplotypes...

...and deciding if they're real

Getting started

Ready to dive in? 🤿 Here's a 30,000-foot view

curl -fsSL https://install.julialang.org | sh -s -- --yes
source ~/.bashrc
julia \
  --startup-file=no \
  --history-file=no \
  --quiet \
  -e 'using Pkg; Pkg.activate(;temp=true); Pkg.add(HapLink)'
echo 'export PATH=$HOME/.julia/bin:$PATH' >> $HOME/.bashrc
source ~/.bashrc

wget https://github.com/ksumngs/HapLink.jl/raw/v1.0.0-rc1/example/reference.fasta
wget https://github.com/ksumngs/HapLink.jl/raw/v1.0.0-rc1/example/sample.bam
wget https://github.com/ksumngs/HapLink.jl/raw/v1.0.0-rc1/example/sample.bam.bai

haplink variants \
  reference.fasta \
  sample.bam \
  --significance 0.5 \
  --depth 1 \
  --quality 10.0 \
  --position 0.01 \
  --frequency 0.05 \
  | tee sample.vcf

haplink consensus \
  reference.fasta \
  sample.vcf \
  | tee consensus.fasta

haplink haplotypes \
  reference.fasta \
  sample.vcf \
  sample.bam \
  --frequency 0.75 \
  | tee sample.yml

haplink sequences \
  reference.fasta \
  sample.yml \
  | tee sample.fasta