VariationCall

HapLink.VariationCall — Type

VariationCall

Represents a Variation that is supported by aligned reads with sufficient metadata to support or refute its validity. It is designed to be converted into a line in Variant Call Format, or a VCF.Record.

Fields

variation::Variation: The Variation of this call

quality::Union{Nothing,Number}: Phred quality of the basecall for variation

filter::Vector{String}: Indicator if variation has passed filters and is actually a variant call, or a list of criteria that have failed it

depth::Union{Nothing,UInt}: The number of reads that cover leftposition(variation)

strandbias::Union{Nothing,Float64}: The fraction of times variation appears on a positive strand

altdepth::Union{Nothing,UInt}: The number of types variation occurs

readpos::Union{Nothing,UInt}: The averagerelative position of variation in each read

pvalue::Union{Nothing,Float64}: The Fisher's Exact Test $p$-value of this call

Getter methods

HapLink.variation — Method

variation(vc::VariationCall) -> SequenceVariation.Variation

Gets the Variation of a VariationCall

HapLink.quality — Method

quality(vc::VariationCall) -> Union{Nothing,Float64}

Gets the average phred quality score of vc, if known. Returns nothing if unknown.

See also quality(::VariationInfo), quality(::VariationPileup)

Missing docstring.

Missing docstring for filter(::VariationCall). Check Documenter's build log for details.

HapLink.depth — Method

depth(vc::VariationCall) -> Union{Nothing,UInt}

Gets the number of times the position of vc appears total. Returns nothing if unknown.

See also depth(::VariationPileup)

HapLink.strand_bias — Method

strand_bias(vc::VariationCall) -> Union{Nothing,Float64}

Gets the fraction of times vc appears on the positive strand. Returns nothing if unknown.

See also strand(::VariationInfo), strand(::VariationPileup)

HapLink.altdepth — Method

altdepth(vc::VariationCall) -> Union{Nothing,UInt}

Gets the number of times vc appears. Returns nothing if unknown.

See also altdepth(::VariationPileup)

HapLink.readpos — Method

readpos(vc::VariationCall) -> Union{Nothing,Float64}

Gets the average relative position of vc. Returns nothing if unknown.

See also readpos(::VariationPileup)

HapLink.p_value — Method

p_value(vc::VariationCall) -> Union{Nothing,Float64}

Gets the $p$-value of the observed statistic of vc. Returns nothing if unknown.

See also variation_test

HapLink.frequency — Method

frequency(vc::VariationCall) -> Float64

Gets the alternate allele frequency of vc

See also frequency(::VariationPileup)

Variant calling methods

HapLink.variation_test — Function

variation_test(depth::Int, altdepth::Int, quality::Float64)

Conducts a Fisher's Exact Test to deterimine the likelihood of a variant with total depth and variation depth altdepth occuring, given an average basecall quality. Returns the $p$-value of the test.

HapLink.call_variant — Function

call_variant(
    pileup::VariationPileup,
    α::Float64;
    D::Union{Nothing,Int}=nothing,
    Q::Union{Nothing,Float64}=nothing,
    X::Union{Nothing,Float64}=nothing,
    F::Union{Nothing,Float64}=nothing,
    S::Union{Nothing,Float64}=nothing,
) -> VariationCall

Calls variant from a pileup.

Arguments

pileup::VariationPileup: The pileup to call variants from

α::Float64: Fisher's Exact Test significance ($α$) level to call variants

Keywords

Note

Leave any keyword undefined to skip filtering based on that field

D::Union{Nothing,Int}=nothing: Minimum total read depth for a variant to be called

Q::Union{Nothing,Float64}=nothing: Minimum average Phred quality for a variant to be called

X::Union{Nothing,Float64}=nothing: Maximum average distance variant can be from read edge to be called

F::Union{Nothing,Float64}=nothing: Minimum alternate frequency for a variant to be called

S::Union{Nothing,Float64}=nothing: Maximum proportion of the number of times variant can appear on one strand versus the other