VariationInfo

HapLink.VariationInfo — Type

VariationInfo{S<:BioSequence,T<:BioSymbol}

Represents statistics associated with a SequenceVariation.Variation within an aligned sequencing read.

Fields

variation::Variation{S,T}: The Variation this object represents

readpos::Float64: The location where variation occurs within a sequencing read

quality::Float64: The phred-scaled basecall quality of variation

Strand::GenomicFeatures.Strand: Which strand variation appears on

Getter methods

HapLink.variation — Method

variation(vi::VariationInfo) -> SequenceVariation.Variation

Gets the Variation of a VariationInfo

HapLink.readpos — Method

readpos(vi::VariationInfo) -> Float64

Gets the position of variation(vi) within a sequencing read

HapLink.quality — Method

quality(vi::VariationInfo) -> Float64

Gets the phred-scaled basecall quality of variation(vi) within a sequencing read

HapLink.strand — Method

strand(vi::VariationInfo) -> GenomicFeatures.Strand

Gets the strand that variation(vi) appears on

Miscellaneous

HapLink.variationinfos — Function

variationinfos(
    query::Union{SAM.Record,BAM.Record}, reference::NucleotideSeq
) -> Vector{VariationInfo}
variationinfos(
    query::Union{AbstractString,AbstractPath},
    reference::Union{AbstractString,AbstractPath}
) -> Vector{VariationInfo}

Calls Variations based on the alignments in query against reference, and returns every variation call found within query as a Vector{VariationInfo}